+++ packages/press-ui/src/packages/common/component-handler/set-data.d.ts
@@ -1,2 +1,3 @@
 export declare function setData(dialog: any, data: any, func?: string): any;
 export declare function vmSet(dialog: any, ...args: Array<any>): void;
+export declare function getVMInVue3(context: any): any;