+++ packages/press-ui/src/packages/common/functional-component/index.d.ts
@@ -1,4 +1,4 @@
-export declare function selectComponent(context: any, selector: string): any;
+export declare function selectComponent(context: any, selector: string | Function): any;
 export declare function showFunctionalComponent(options: {
   context?: any;
   selector: string;