+++ packages/press-ui/src/packages/common/utils/system.js
@@ -1,23 +1,7 @@
 import { canIUseGroupSetData } from './version';
 import { isNotInUni } from './utils';
 export { nextTick } from '../vue3/adapter';
-
-let systemInfo = {};
-
-export function getSystemInfoSync() {
-  try {
-    if (!Object.keys(systemInfo).length
-      && typeof uni !== 'undefined'
-      && typeof uni?.getSystemInfoSync === 'function'
-    ) {
-      systemInfo = uni.getSystemInfoSync();
-    }
-  } catch (err) {
-    console.log('[getSystemInfoSync] err', err);
-  }
-
-  return systemInfo;
-}
+export { getSystemInfoSync } from './version';
 
 
 export function requestAnimationFrame(cb) {