+++ packages/press-ui/src/packages/press-dialog/handler.d.ts
@@ -2,6 +2,9 @@ interface IDialog {
   (options: any): any;
   show: (options: any) => Promise<any>;
   clear: () => any;
+
+  setDefaultOptions: (options: any) => void;
+  resetDefaultOptions: () => void;
 }
 const Dialog: IDialog;