+++ packages/press-ui/src/packages/press-fab/demo.vue
@@ -37,7 +37,7 @@
 <script>
 import PressFab from 'press-ui/press-fab/press-fab.vue';
 import PressCell from 'press-ui/press-cell/press-cell.vue';
-import { isNotInUni } from 'press-ui/common/utils/utils';
+// import { isNotInUni } from 'press-ui/common/utils/utils';
 import { forceUpdate } from 'press-ui/common/vue3/adapter';
 import { getWindowWidth } from 'press-ui/common/dom/rect';