+++ packages/press-ui/src/packages/press-toast/handler.d.ts
@@ -4,6 +4,9 @@ interface IToast {
   success: (options: any) => any;
   fail: (options: any) => any;
   clear: () => any;
+
+  setDefaultOptions: (options: any) => void;
+  resetDefaultOptions: () => void;
 }
 const Toast: IToast;